ハッシュ検証を実装

+++ web/prebuild.ts
@@ -28,20 +28,22 @@ if (!res.ok) {
 	console.error(res.error);
 	throw new Error();
 }
-if(!res.payload){
-	console.error("payload not found")
-	throw new Error()
+if (!res.payload) {
+	console.error("payload not found");
+	throw new Error();
 }
-const hash=crypto.hash("SHA256",JSON.stringify(res.payload),"base64")
-const oldhash=""
-
-const metadata = { lastupdate: new Date(),hash };
+const hash = crypto.hash("SHA256", JSON.stringify(res.payload), "base64");
+const {hash:oldhash} = await fetch("https://polaris-consts.pages.dev/data/metadata.json").then(
+	(r) => r.json() as Promise<{ hash: string }>,
+);
+if(hash===oldhash){
+	process.exit(11)
+}
+const metadata = { lastupdate: new Date(), hash };
 const datadir = path.join(import.meta.dirname, "public", "data");
 try {
 	await fs.rm(datadir, { force: true, recursive: true });
 } catch (e) {}
-await fs.mkdir(path.join(import.meta.dirname, "public", "data"),{recursive:true});
+await fs.mkdir(path.join(import.meta.dirname, "public", "data"), { recursive: true });
 await fs.writeFile(path.join(datadir, "metadata.json"), JSON.stringify(metadata, undefined, 0));
 await fs.writeFile(path.join(datadir, "data.json"), JSON.stringify(res.payload, undefined, 0));
-
-process.exitCode=0
